chore(merkle_root): split function to it's own file
